pyproject.toml
[project]
name = "gerenciamento"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "VitorAndreSilva", email = "alunovitorandre@gmail.com"},
]
dependencies = ["django>=6.0", "djangorestframework>=3.16.1", "djangorestframework-simplejwt>=5.5.1", "django-filter>=25.2"]
requires-python = "==3.13.*"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = false
